+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Minor: Use notifier-specific screenshot filenames. (#79)
+- Minor: Support notifier-specific webhook URLs. (#78)
 - Dev: Add mockito test suite for notifiers. (#74, #75)
 - Dev: Utilize more dependency injection. (#73)